Updates to data API object.
